Address GitLab issue #217: Warnings on OpenBSD 6.0. Prevent possible format warning is suseconds_t != long on a platform.

+++ ntpd/refclock_shm.c
@@ -520,7 +520,7 @@ shm_timer(
 	 */
 	/* a_lastcode is seen as timecode with: ntpq -c cv [associd] */
 	c = snprintf(pp->a_lastcode, sizeof(pp->a_lastcode), "%ld.%09ld",
-		     (long)shm_stat.tvt.tv_sec, shm_stat.tvt.tv_nsec);
+		     (long)shm_stat.tvt.tv_sec, (long)shm_stat.tvt.tv_nsec);
 	pp->lencode = (c < (int)sizeof(pp->a_lastcode)) ? c : 0;
 
 	/* check 1: age control of local time stamp */
